test: move the auditor conformance suite under a collected test directory
The suite landed at tests/conformance/, a top-level directory no CI job
hands to pytest: the sharded Test jobs run scripts/run_tests.py, whose
default test directory is tests/unit, and no workflow names
tests/conformance/. scripts/check_test_collection.py exists to catch
exactly this and failed the build, because a suite nothing collects
cannot fail one either — the scoreboard would have reported a score no
regression could ever move.

Move it to tests/integration/conformance/auditor/, which the Integration
tests job collects via --test-dir tests/integration and which the impact
analyser's TEST_DIRS already covers, so a pass→fail regression in a
vector now breaks the build. The move is mechanical: module paths, the
scoreboard's SUITE_PATH, the marker description and the repo-root depth
in isolation.py follow the new location. tests/AGENTS.md goes back to its
original text, since the suite is no longer a top-level tree.

Also add the release-notes fragment for the one user-visible change in
the branch: the standalone receipt verifier's script entry point passed
main to sys.exit instead of calling it.
