+++ pyproject.toml
@@ -807,7 +807,7 @@ markers = [
     "stress: resource-leak / memory-growth / deadlock stress tests (TC-C); gated off by default CI, run nightly via `-m stress`",
     "allow_network: opt out of the autouse unit no-network guard (rare; a genuine integration test belongs in tests/integration/ instead)",
     "protocol: protocol compatibility tests; applied to every item under tests/protocol/ by that directory's conftest, and run by the publish workflow's release gate",
-    "auditor_question: binds one conformance vector under tests/conformance/auditor/ to the question it answers; the argument is a question number registered in that suite's questions.py",
+    "auditor_question: binds one conformance vector under tests/integration/conformance/auditor/ to the question it answers; the argument is a question number registered in that suite's questions.py",
 ]
 # Visible warnings without converting to errors; individual tests or runs
 # can tighten via `pytest -W error::DeprecationWarning`.
