+++ tests/AGENTS.md
@@ -2,8 +2,8 @@
 
 Layered pytest suite. `unit/` is the big one (2400+ files, no network);
 `integration/` needs a running server; `contract/` holds the adapter
-capability contracts; plus `conformance/`, `property/`, `snapshot/`,
-`golden/`, `perf/`, `chaos/`, `stress/`, `pentest/`, and `benchmarks/`.
+capability contracts; plus `property/`, `snapshot/`, `golden/`,
+`perf/`, `chaos/`, `stress/`, `pentest/`, and `benchmarks/`.
 
 ## How to run
